api/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
argcomplete==3.6.1
bcrypt==4.3.0
blinker==1.9.0
boto3==1.37.24
botocore==1.37.24
cachetools==5.5.1
certifi==2025.1.31
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
durationpy==0.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-auth==2.38.0
google-genai==1.2.0
hjson==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kappa==0.6.0
Mako==1.3.9
MarkupSafe==3.0.2
placebo==0.9.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
SQLAlchemy==2.0.36
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
zappa==0.59.0
